more debug stuff

+++ solace-openflowj/src/main/scala/com/yevhenii/solace/metrics/MetricReporter.scala
@@ -36,7 +36,7 @@ class MetricReporter(config: Config, containerId: String = "localhost") extends 
     asMap.foreach {
       case (IncomingOF, msg) => statsdClient.increment(s"of_in_$msg", s"dpid:$dpid")
       case (ResultOF, msg) => statsdClient.increment(s"of_out_$msg", s"dpid:$dpid")
-      case (SizeOF, size) => statsdClient.count("of_size", size.asInstanceOf[Long], s"dpid:$dpid", s"sender:$sender")
+      case (SizeOF, size) => statsdClient.count("of_size", size.asInstanceOf[Int], s"dpid:$dpid", s"sender:$sender")
       case (ProcessingTime, time) => statsdClient.count("of_in_time", time.asInstanceOf[Long], s"dpid:$dpid", s"type:$inMsgType")
       case (key, value) =>
     }
